+++ src/nglui/__init__.py
@@ -1,5 +1,6 @@
 from nglui.easyviewer import EasyViewer, annotation
 from . import statebuilder
 from . import nglite
+from . import parser
 
 __version__ = "2.2.0"